Introduce ETH staking limit
We now have an ETH staking limit instead of a number of
validators. The frontend should show an error in the staking page and
in the add/edit validator view when the appropriate status code is
returned with the information of how much ETH is staked and how much
the limit is.

Solves https://github.com/orgs/rotki/projects/11/views/3?pane=issue&itemId=120386420