Change how gnosis pay and monerium tx get their metadata
This solves an issue where sometimes the metadata wasn't added to the
notes even when the information was in the database.

What it does is to add a new type of functions that run after the decoding
and what they do is to use local information or query apis to fetch metadata.
The reason to do it in a new step is so we can batch the queries and instead
of querying one by one we can get all the information in one api response from
the remote seerver.

Also this honors the premium status and had to make the decoder class aware
of it.

Co-authored-by: Yabir Benchakhtir <git@yabirgb.com>
Co-authored-by: Nicholas Yoder <nicholasyoder@proton.me>