Fix database table is locked during upload
Fixes #5038 again. Adding a new test for it too.

wal_checkpoint (irrespective of type) will fail with database table is
locked if there is any readers on that DB that execute a progress
handler.

So the solution is to acquire the progress handler lock for the DB for
the duration of the checkpointing.

Abstracted that logic in a function.