Improve the ignored_assets query in get_tokens_for_address
Using NOT EXISTS instead of NOT IN the subquery since it is a faster
way to do the same thing.

 1. NOT EXISTS stops at the first match instead of building the entire
 subquery result set
 2. The correlated subquery approach avoids the large NOT IN
 comparison

Added logs to time it in the normnal balance query. For me.

NOT EXISTS query took 0.0001s - 0.0003s across the entire run. So
pretty steady results.

NOT IN query took 0.0001s - 1.4870s  (average was 0.002-0.004). So not
steady and some cases were super slow.