Remove FREE_ETH_TX_LIMIT
We had different limits in the past but we checked the logic around this one and
it seems to be used only in `ensure_tx_data_exists` and `ensure_genesis_tx_data_exists`.

After discussing it in discord we said to remove it since it is not really limiting
the user. We keep `FREE_HISTORY_EVENTS_LIMIT` but we mentioned that we want to change
it along the tiers.