Skip v1 to v2 monerium migrations
Introduces a new condition to clear the processed events and stop iterating.
Made the check of the transaction hashes into a function to avoid having
the string in memory all the time.

It also introduces a mechanism to stop processing the transactions. Since those
can have thousend of events and each one is processed I believe is a good idea
to stop processing the transactions. Also I believe this is a thing that could
be used in other scenarios like spam transactions.