Update the list of 0x settlers
Adds missing entries of the 0x settler contract in the supported chains.
Found that the contract `0x00000000000004533fe15556b1e086bb1a72ceae` is an NFT
that they use to track the deployments. When a new deployment happens the old settler
transfers the NFT to the new one. Using this contract I tracked the deployments
that we were missing.

Also notice that before the deployments appeared labeled in etherscan with versions but
is not the case anymore, not sure if they stopped doing it.

Closes https://github.com/orgs/rotki/projects/11/views/3?pane=issue&itemId=130550124