Ensure that FK are always set for globaldb upgrades
THe problem was that if we had assets updates pending
(needed to not miss them before upgrading) they were done
and the foreign keys got activated. If instead a user just upgraded and
didn't had to pull remote data the foreign keys weren't set until
after the globaldb upgrades. That broke assumptions in the tests as seen now.