Remove dead code bug
I think that the condition was always true as event_type with
`HistoryEventSubtype` comparison would always be false.

What's more i am not sure what the original intention was as we could
not return a data without an `event_type`.

Added a comment to explain that at serialization we knowingly use
subtype in the data place of event type to match the api's expectations