Correctly query CRV balances
What was happening is that the lock event was being detected along the gauge
deposit events and the veCRV token was treated as a gauge. As such we were calling
balanceOf in the veCRV and it returns the voting power instead of the total CRV locked.

The new logic excludes the CRV token query from the gauges logic and calls `locked` instead
to get the actual CRV in the protocol