Fix process_result() tuple serialization
The tuple serialization at process_result just turned them into a
list. Which would explode in our faces with type errors if any non
primitive type was in there and we tried to export via API.

This never hit us so far since in all the places where tuples are
exported we convert them manually.

But that's a footgun we should fix.

(Both claude and codex warn about this)