Allow editing a tag name
Fix #11021

Backend: Did it in a bit more manual way without changing the DB schema. WE
could do a DB upgrade and change the schema and simplify the editing
here.

Frontend: Made it so that clicking a Tag in the evm account view opens
the tag editor. Also made it so that clicking save after that moves
you back to the evm account view.